requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit
transformers==4.36.2
torch==2.7.1
jiwer
pillow
requests
openai-whisper
ffmpeg-python
streamlit-mic-recorder 